Refactoring, change progress bar style, add common getResponse function
