+++ app/src/main/java/com/innoprog/android/feature/feed/newsfeed/domain/models/QueryPage.kt
@@ -9,5 +9,3 @@ data class QueryPage(
     val projectId: String? = null,
     val authorId: String? = null
 )
-
-
