+++ app/src/main/java/com/innoprog/android/feature/profile/profiledetails/data/network/ProfileRetrofitClient.kt
@@ -98,7 +98,6 @@ class ProfileRetrofitClient @Inject constructor(
     }
 
     companion object {
-
         private const val PAGE_SIZE = 50
     }
 }
